@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
     babel.lists
     ~~~~~~~~~~~
@@ -11,7 +10,7 @@
      * ``LC_ALL``, and
      * ``LANG``
 
-    :copyright: (c) 2015-2021 by the Babel Team.
+    :copyright: (c) 2015-2022 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
@@ -69,11 +68,10 @@ def format_list(lst, style='standard', locale=DEFAULT_LOCALE):
         return lst[0]
 
     if style not in locale.list_patterns:
-        raise ValueError('Locale %s does not support list formatting style %r (supported are %s)' % (
-            locale,
-            style,
-            list(sorted(locale.list_patterns)),
-        ))
+        raise ValueError(
+            f'Locale {locale} does not support list formatting style {style!r} '
+            f'(supported are {sorted(locale.list_patterns)})'
+        )
     patterns = locale.list_patterns[style]
 
     if len(lst) == 2:
