@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
     babel.localedata
     ~~~~~~~~~~~~~~~~
@@ -8,17 +7,17 @@
     :note: The `Locale` class, which uses this module under the hood, provides a
            more convenient interface for accessing the locale data.
 
-    :copyright: (c) 2013-2021 by the Babel Team.
+    :copyright: (c) 2013-2022 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
+import pickle
 import os
 import re
 import sys
 import threading
+from collections import abc
 from itertools import chain
-
-from babel._compat import pickle, string_types, abc
 
 
 _cache = {}
@@ -33,7 +32,7 @@ def normalize_locale(name):
     Returns the normalized locale ID string or `None` if the ID is not
     recognized.
     """
-    if not name or not isinstance(name, string_types):
+    if not name or not isinstance(name, str):
         return None
     name = name.strip().lower()
     for locale_id in chain.from_iterable([_cache, locale_identifiers()]):
@@ -51,10 +50,10 @@ def resolve_locale_filename(name):
 
     # Ensure we're not left with one of the Windows reserved names.
     if sys.platform == "win32" and _windows_reserved_name_re.match(os.path.splitext(name)[0]):
-        raise ValueError("Name %s is invalid on Windows" % name)
+        raise ValueError(f"Name {name} is invalid on Windows")
 
     # Build the path.
-    return os.path.join(_dirname, '%s.dat' % name)
+    return os.path.join(_dirname, f"{name}.dat")
 
 
 def exists(name):
@@ -64,7 +63,7 @@ def exists(name):
 
     :param name: the locale identifier string
     """
-    if not name or not isinstance(name, string_types):
+    if not name or not isinstance(name, str):
         return False
     if name in _cache:
         return True
@@ -184,7 +183,7 @@ def merge(dict1, dict2):
             dict1[key] = val1
 
 
-class Alias(object):
+class Alias:
     """Representation of an alias in the locale data.
 
     An alias is a value that refers to some other part of the locale data,
@@ -195,7 +194,7 @@ class Alias(object):
         self.keys = tuple(keys)
 
     def __repr__(self):
-        return '<%s %r>' % (type(self).__name__, self.keys)
+        return f"<{type(self).__name__} {self.keys!r}>"
 
     def resolve(self, data):
         """Resolve the alias based on the given data.
