@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
     babel.messages.checkers
     ~~~~~~~~~~~~~~~~~~~~~~~
@@ -7,12 +6,11 @@
 
     :since: version 0.9
 
-    :copyright: (c) 2013-2021 by the Babel Team.
+    :copyright: (c) 2013-2022 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
 from babel.messages.catalog import TranslationError, PYTHON_FORMAT
-from babel._compat import string_types, izip
 
 
 #: list of format chars that are compatible to each other
@@ -26,7 +24,7 @@ _string_format_compatibilities = [
 def num_plurals(catalog, message):
     """Verify the number of plurals in the translation."""
     if not message.pluralizable:
-        if not isinstance(message.string, string_types):
+        if not isinstance(message.string, str):
             raise TranslationError("Found plural forms for non-pluralizable "
                                    "message")
         return
@@ -54,7 +52,7 @@ def python_format(catalog, message):
     if not isinstance(msgstrs, (list, tuple)):
         msgstrs = (msgstrs,)
 
-    for msgid, msgstr in izip(msgids, msgstrs):
+    for msgid, msgstr in zip(msgids, msgstrs):
         if msgstr:
             _validate_format(msgid, msgstr)
 
@@ -134,7 +132,7 @@ def _validate_format(format, alternative):
         if len(a) != len(b):
             raise TranslationError('positional format placeholders are '
                                    'unbalanced')
-        for idx, ((_, first), (_, second)) in enumerate(izip(a, b)):
+        for idx, ((_, first), (_, second)) in enumerate(zip(a, b)):
             if not _compatible(first, second):
                 raise TranslationError('incompatible format for placeholder '
                                        '%d: %r and %r are not compatible' %
@@ -146,7 +144,7 @@ def _validate_format(format, alternative):
         type_map = dict(a)
         for name, typechar in b:
             if name not in type_map:
-                raise TranslationError('unknown named placeholder %r' % name)
+                raise TranslationError(f'unknown named placeholder {name!r}')
             elif not _compatible(typechar, type_map[name]):
                 raise TranslationError('incompatible format for '
                                        'placeholder %r: '
