@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
     babel.support
     ~~~~~~~~~~~~~
@@ -8,7 +7,7 @@
 
     .. note: the code in this module is not used by Babel itself
 
-    :copyright: (c) 2013-2021 by the Babel Team.
+    :copyright: (c) 2013-2022 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
@@ -18,12 +17,11 @@ import locale
 from babel.core import Locale
 from babel.dates import format_date, format_datetime, format_time, \
     format_timedelta
-from babel.numbers import format_number, format_decimal, format_currency, \
-    format_percent, format_scientific
-from babel._compat import PY2, text_type, text_to_native
+from babel.numbers import format_decimal, format_currency, format_compact_currency, \
+    format_percent, format_scientific, format_compact_decimal
 
 
-class Format(object):
+class Format:
     """Wrapper class providing the various date and number formatting functions
     bound to a specific locale and time-zone.
 
@@ -99,7 +97,7 @@ class Format(object):
         >>> fmt.number(1099)
         u'1,099'
         """
-        return format_number(number, locale=self.locale)
+        return format_decimal(number, locale=self.locale)
 
     def decimal(self, number, format=None):
         """Return a decimal number formatted for the locale.
@@ -110,10 +108,28 @@ class Format(object):
         """
         return format_decimal(number, format, locale=self.locale)
 
+    def compact_decimal(self, number, format_type='short', fraction_digits=0):
+        """Return a number formatted in compact form for the locale.
+
+        >>> fmt = Format('en_US')
+        >>> fmt.compact_decimal(123456789)
+        u'123M'
+        """
+        return format_compact_decimal(number, format_type=format_type,
+                                      fraction_digits=fraction_digits,
+                                      locale=self.locale)
+
     def currency(self, number, currency):
         """Return a number in the given currency formatted for the locale.
         """
         return format_currency(number, currency, locale=self.locale)
+
+    def compact_currency(self, number, currency, format_type='short', fraction_digits=0):
+        """Return a number in the given currency formatted for the locale
+        using the compact number format.
+        """
+        return format_compact_currency(number, currency, format_type=format_type,
+                                        fraction_digits=fraction_digits, locale=self.locale)
 
     def percent(self, number, format=None):
         """Return a number formatted as percentage for the locale.
@@ -130,7 +146,7 @@ class Format(object):
         return format_scientific(number, locale=self.locale)
 
 
-class LazyProxy(object):
+class LazyProxy:
     """Class for proxy objects that delegate to a specified function to evaluate
     the actual object.
 
@@ -194,7 +210,7 @@ class LazyProxy(object):
     def __contains__(self, key):
         return key in self.value
 
-    def __nonzero__(self):
+    def __bool__(self):
         return bool(self.value)
 
     def __dir__(self):
@@ -208,9 +224,6 @@ class LazyProxy(object):
 
     def __str__(self):
         return str(self.value)
-
-    def __unicode__(self):
-        return unicode(self.value)
 
     def __add__(self, other):
         return self.value + other
@@ -289,7 +302,7 @@ class LazyProxy(object):
         )
 
 
-class NullTranslations(gettext.NullTranslations, object):
+class NullTranslations(gettext.NullTranslations):
 
     DEFAULT_DOMAIN = None
 
@@ -305,7 +318,7 @@ class NullTranslations(gettext.NullTranslations, object):
         # some *gettext methods (including '.gettext()') rely on the attributes.
         self._catalog = {}
         self.plural = lambda n: int(n != 1)
-        super(NullTranslations, self).__init__(fp=fp)
+        super().__init__(fp=fp)
         self.files = list(filter(None, [getattr(fp, 'name', None)]))
         self.domain = self.DEFAULT_DOMAIN
         self._domains = {}
@@ -320,6 +333,9 @@ class NullTranslations(gettext.NullTranslations, object):
         """Like ``lgettext()``, but look the message up in the specified
         domain.
         """
+        import warnings
+        warnings.warn('ldgettext() is deprecated, use dgettext() instead',
+                      DeprecationWarning, 2)
         return self._domains.get(domain, self).lgettext(message)
 
     def udgettext(self, domain, message):
@@ -340,6 +356,9 @@ class NullTranslations(gettext.NullTranslations, object):
         """Like ``lngettext()``, but look the message up in the specified
         domain.
         """
+        import warnings
+        warnings.warn('ldngettext() is deprecated, use dngettext() instead',
+                      DeprecationWarning, 2)
         return self._domains.get(domain, self).lngettext(singular, plural, num)
 
     def udngettext(self, domain, singular, plural, num):
@@ -372,11 +391,6 @@ class NullTranslations(gettext.NullTranslations, object):
             if self._fallback:
                 return self._fallback.pgettext(context, message)
             return message
-        # Encode the Unicode tmsg back to an 8-bit string, if possible
-        if self._output_charset:
-            return text_to_native(tmsg, self._output_charset)
-        elif self._charset:
-            return text_to_native(tmsg, self._charset)
         return tmsg
 
     def lpgettext(self, context, message):
@@ -384,16 +398,12 @@ class NullTranslations(gettext.NullTranslations, object):
         preferred system encoding, if no other encoding was explicitly set with
         ``bind_textdomain_codeset()``.
         """
-        ctxt_msg_id = self.CONTEXT_ENCODING % (context, message)
-        missing = object()
-        tmsg = self._catalog.get(ctxt_msg_id, missing)
-        if tmsg is missing:
-            if self._fallback:
-                return self._fallback.lpgettext(context, message)
-            return message
-        if self._output_charset:
-            return tmsg.encode(self._output_charset)
-        return tmsg.encode(locale.getpreferredencoding())
+        import warnings
+        warnings.warn('lpgettext() is deprecated, use pgettext() instead',
+                      DeprecationWarning, 2)
+        tmsg = self.pgettext(context, message)
+        encoding = getattr(self, "_output_charset", None) or locale.getpreferredencoding()
+        return tmsg.encode(encoding)
 
     def npgettext(self, context, singular, plural, num):
         """Do a plural-forms lookup of a message id.  `singular` is used as the
@@ -409,10 +419,6 @@ class NullTranslations(gettext.NullTranslations, object):
         ctxt_msg_id = self.CONTEXT_ENCODING % (context, singular)
         try:
             tmsg = self._catalog[(ctxt_msg_id, self.plural(num))]
-            if self._output_charset:
-                return text_to_native(tmsg, self._output_charset)
-            elif self._charset:
-                return text_to_native(tmsg, self._charset)
             return tmsg
         except KeyError:
             if self._fallback:
@@ -427,12 +433,14 @@ class NullTranslations(gettext.NullTranslations, object):
         preferred system encoding, if no other encoding was explicitly set with
         ``bind_textdomain_codeset()``.
         """
+        import warnings
+        warnings.warn('lnpgettext() is deprecated, use npgettext() instead',
+                      DeprecationWarning, 2)
         ctxt_msg_id = self.CONTEXT_ENCODING % (context, singular)
         try:
             tmsg = self._catalog[(ctxt_msg_id, self.plural(num))]
-            if self._output_charset:
-                return tmsg.encode(self._output_charset)
-            return tmsg.encode(locale.getpreferredencoding())
+            encoding = getattr(self, "_output_charset", None) or locale.getpreferredencoding()
+            return tmsg.encode(encoding)
         except KeyError:
             if self._fallback:
                 return self._fallback.lnpgettext(context, singular, plural, num)
@@ -454,7 +462,7 @@ class NullTranslations(gettext.NullTranslations, object):
         if tmsg is missing:
             if self._fallback:
                 return self._fallback.upgettext(context, message)
-            return text_type(message)
+            return str(message)
         return tmsg
 
     def unpgettext(self, context, singular, plural, num):
@@ -475,9 +483,9 @@ class NullTranslations(gettext.NullTranslations, object):
             if self._fallback:
                 return self._fallback.unpgettext(context, singular, plural, num)
             if num == 1:
-                tmsg = text_type(singular)
+                tmsg = str(singular)
             else:
-                tmsg = text_type(plural)
+                tmsg = str(plural)
         return tmsg
 
     def dpgettext(self, domain, context, message):
@@ -525,9 +533,8 @@ class NullTranslations(gettext.NullTranslations, object):
         return self._domains.get(domain, self).lnpgettext(context, singular,
                                                           plural, num)
 
-    if not PY2:
-        ugettext = gettext.NullTranslations.gettext
-        ungettext = gettext.NullTranslations.ngettext
+    ugettext = gettext.NullTranslations.gettext
+    ungettext = gettext.NullTranslations.ngettext
 
 
 class Translations(NullTranslations, gettext.GNUTranslations):
@@ -541,12 +548,11 @@ class Translations(NullTranslations, gettext.GNUTranslations):
         :param fp: the file-like object the translation should be read from
         :param domain: the message domain (default: 'messages')
         """
-        super(Translations, self).__init__(fp=fp)
+        super().__init__(fp=fp)
         self.domain = domain or self.DEFAULT_DOMAIN
 
-    if not PY2:
-        ugettext = gettext.GNUTranslations.gettext
-        ungettext = gettext.GNUTranslations.ngettext
+    ugettext = gettext.GNUTranslations.gettext
+    ungettext = gettext.GNUTranslations.ngettext
 
     @classmethod
     def load(cls, dirname=None, locales=None, domain=None):
@@ -571,8 +577,8 @@ class Translations(NullTranslations, gettext.GNUTranslations):
             return cls(fp=fp, domain=domain)
 
     def __repr__(self):
-        return '<%s: "%s">' % (type(self).__name__,
-                               self._info.get('project-id-version'))
+        version = self._info.get('project-id-version')
+        return f'<{type(self).__name__}: "{version}">'
 
     def add(self, translations, merge=True):
         """Add the given translations to the catalog.
