@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 #
-# Copyright (C) 2007-2011 Edgewall Software, 2013-2021 the Babel team
+# Copyright (C) 2007-2011 Edgewall Software, 2013-2022 the Babel team
 # All rights reserved.
 #
 # This software is licensed as described in the file LICENSE, which
@@ -13,9 +12,9 @@
 
 import os
 import unittest
+from io import BytesIO
 
 from babel.messages import mofile, Catalog
-from babel._compat import BytesIO, text_type
 from babel.support import Translations
 
 
@@ -29,12 +28,11 @@ class ReadMoTestCase(unittest.TestCase):
                                'LC_MESSAGES', 'messages.mo')
         with open(mo_path, 'rb') as mo_file:
             catalog = mofile.read_mo(mo_file)
-            self.assertEqual(2, len(catalog))
-            self.assertEqual('TestProject', catalog.project)
-            self.assertEqual('0.1', catalog.version)
-            self.assertEqual('Stange', catalog['bar'].string)
-            self.assertEqual(['Fuhstange', 'Fuhstangen'],
-                             catalog['foobar'].string)
+            assert len(catalog) == 2
+            assert catalog.project == 'TestProject'
+            assert catalog.version == '0.1'
+            assert catalog['bar'].string == 'Stange'
+            assert catalog['foobar'].string == ['Fuhstange', 'Fuhstangen']
 
 
 class WriteMoTestCase(unittest.TestCase):
@@ -55,16 +53,11 @@ class WriteMoTestCase(unittest.TestCase):
         mofile.write_mo(buf, catalog)
         buf.seek(0)
         translations = Translations(fp=buf)
-        self.assertEqual(u'Voh', translations.ugettext('foo'))
-        assert isinstance(translations.ugettext('foo'), text_type)
-        self.assertEqual(u'Es gibt', translations.ungettext('There is', 'There are', 1))
-        assert isinstance(translations.ungettext('There is', 'There are', 1), text_type)
-        self.assertEqual(u'Fizz', translations.ugettext('Fizz'))
-        assert isinstance(translations.ugettext('Fizz'), text_type)
-        self.assertEqual(u'Fuzz', translations.ugettext('Fuzz'))
-        assert isinstance(translations.ugettext('Fuzz'), text_type)
-        self.assertEqual(u'Fuzzes', translations.ugettext('Fuzzes'))
-        assert isinstance(translations.ugettext('Fuzzes'), text_type)
+        assert translations.ugettext('foo') == 'Voh'
+        assert translations.ungettext('There is', 'There are', 1) == 'Es gibt'
+        assert translations.ugettext('Fizz') == 'Fizz'
+        assert translations.ugettext('Fuzz') == 'Fuzz'
+        assert translations.ugettext('Fuzzes') == 'Fuzzes'
 
     def test_more_plural_forms(self):
         catalog2 = Catalog(locale='ru_RU')
@@ -93,4 +86,4 @@ class WriteMoTestCase(unittest.TestCase):
         translations = Translations(fp=buf1)
         translations.add_fallback(Translations(fp=buf2))
 
-        self.assertEqual(u'Flou', translations.ugettext('Fuzz'))
+        assert translations.ugettext('Fuzz') == 'Flou'
