@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
     babel.localtime
     ~~~~~~~~~~~~~~~
@@ -6,7 +5,7 @@
     Babel specific fork of tzlocal to determine the local timezone
     of the system.
 
-    :copyright: (c) 2013-2021 by the Babel Team.
+    :copyright: (c) 2013-2022 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
