@@ -1,11 +1,10 @@
-# -*- coding: utf-8 -*-
 """
     babel.messages
     ~~~~~~~~~~~~~~
 
     Support for ``gettext`` message catalogs.
 
-    :copyright: (c) 2013-2021 by the Babel Team.
+    :copyright: (c) 2013-2022 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
