@@ -1,11 +1,10 @@
-# -*- coding: utf-8 -*-
 """
     babel.messages.plurals
     ~~~~~~~~~~~~~~~~~~~~~~
 
     Plural form definitions.
 
-    :copyright: (c) 2013-2021 by the Babel Team.
+    :copyright: (c) 2013-2022 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
@@ -154,7 +153,7 @@ PLURALS = {
     # Latvian
     'lv': (3, '(n%10==1 && n%100!=11 ? 0 : n != 0 ? 1 : 2)'),
     # Maltese - From Pootle's PO's
-    'mt': (4, '(n==1 ? 0 : n==0 || ( n%100=>1 && n%100<=10) ? 1 : (n%100>10 && n%100<20 ) ? 2 : 3)'),
+    'mt': (4, '(n==1 ? 0 : n==0 || ( n%100>=1 && n%100<=10) ? 1 : (n%100>10 && n%100<20 ) ? 2 : 3)'),
     # Norwegian Bokmål
     'nb': (2, '(n != 1)'),
     # Dutch
