@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # Babel documentation build configuration file, created by
 # sphinx-quickstart on Wed Jul  3 17:53:01 2013.
@@ -44,16 +43,16 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'Babel'
-copyright = u'2021, The Babel Team'
+copyright = u'2022, The Babel Team'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
 # The short X.Y version.
-version = '2.9'
+version = '2.11'
 # The full version, including alpha/beta/rc tags.
-release = '2.9.1'
+release = '2.11.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -224,8 +223,7 @@ latex_logo = '_static/logo.png'
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    ('index_', 'babel', u'Babel Documentation',
-     [u'The Babel Team'], 1)
+    ('index', 'babel', u'Babel Documentation', [u'The Babel Team'], 1),
 ]
 
 # If true, show URL addresses after external links.
@@ -253,10 +251,10 @@ texinfo_documents = [
 #texinfo_show_urls = 'footnote'
 
 intersphinx_mapping = {
-    'https://docs.python.org/2/': None,
+    'https://docs.python.org/3/': None,
 }
 
 extlinks = {
-    'gh': ('https://github.com/python-babel/babel/issues/%s', '#'),
-    'trac': ('http://babel.edgewall.org/ticket/%s', 'ticket #'),
+    'gh': ('https://github.com/python-babel/babel/issues/%s', '#%s'),
+    'trac': ('http://babel.edgewall.org/ticket/%s', 'ticket #%s'),
 }
