@@ -1,4 +1,3 @@
-# -- encoding: UTF-8 --
 from datetime import time
 
 import babel.dates as dates
@@ -10,6 +9,9 @@ import pytest
     ("de", time(3, 11), "night1"),  # (after, before)
     ("fi", time(0), "midnight"),  # (at)
     ("en_US", time(12), "noon"),  # (at)
+    ("en_US", time(21), "night1"),  # (from, before) across 0:00
+    ("en_US", time(5), "night1"),  # (from, before) across 0:00
+    ("en_US", time(6), "morning1"),  # (from, before)
     ("agq", time(10), "am"),  # no periods defined
     ("agq", time(22), "pm"),  # no periods defined
     ("am", time(14), "afternoon1"),  # (before, after)
